My 2nd submission of this project.
There are two reasons for this 2nd submission:
The 1st reason is to improve my readme file. My instruction from the reviewer is to add sections that explain how to run my project successfully, how the game works, and, how to play the game. I used "Markdown Dingus" and other resources to learn this technology.
The 2nd reason for this submission from the reviewer is: "Your code looks pretty good. You have a few JavaScript syntax errors in app.js. If you run your code through JSHint you can catch tiny errors like these. Another option is to add a linter plugin to your text editor." By the way, the JSLint is a great tool and i will be using it and other 'code correction' technologies from now on. It really does help me to write better code and to understand my mistakes. Thanks for the advice!